### 2016-05-12 #18 @whedon
Submission: Xenomapper - short read mapping in a mixed species context
**Submitting author:** @genomematt (<a href="http://orcid.org/0000-0001-6624-4698">Matthew J Wakefield</a>)
**Repository:** <a href="http://github.com/genomematt/xenomapper/" target ="_blank">http://github.com/genomematt/xenomapper/</a>
**Version:** v1.0.1
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.51798" target ="_blank">10.5281/zenodo.51798 </a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.51337">software archive</a> resolve?

- [x] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00018.pdf](https://github.com/openjournals/joss-reviews/files/261671/10.21105.joss.00018.pdf)
