### 2016-05-16 #18 @xuanxu
## General checks
all checked

## Functionality
all checked
- I confirmed the functionality using code from the test suite

## Documentation
some checked

### Missing:
- Example Usage: There is no example usage or command to run the software using some of the sample data included. **Recommendation**: Add a 'Example usage' section with step-by-step commands to run Xenomapper with files from `xenomapper/tests/data` as input and explanation of each problem solved. More than a single example would be great to have examples of some of the multiple features.

- Community guidelines: There is no mention on how to contribute, report bugs or get support. **Recommendation**: Add a 'How to contribute' section with at least a brief statement encouraging posible contributors to help pointing them to open a new github issue. Even better would be to link to a CONTRIBUTING.md file with extended instructions. Add also comments on how to report a problem (opening an issue should be OK) and seek support (ideas: mailing list, contact email, github issue with a specific tag).

## Software paper:
all checked
- Authors and statement of needs OK
- All DOIs verified

## Nice-to-have improvements:
- License file is included in the repository, but users need to open it to know what is the actual license used. I'd recommend adding a mention in the readme file. Just something like "Licensed under GPL v3" would be useful.

- Tests are included, so that point is checked. But It would be good if instructions on how to run the tests are clearly stated in the README. Actually the command to do it (`python3 -m xenomapper.tests.test_all`) is there, but under the `Alternate Installation` section.

## Recommendation:
I recommend publication after some usage examples and contributing instructions are added to the documentation.
