### 2016-05-16 #18 @arfon
> I recommend publication after some minor revisions: usage examples and contributing instructions should be added to the documentation.

Many thanks for the review @xuanxu. @genomematt - please let us know when you've addressed @xuanxu's comments and we can move forward.
