### 2016-05-21 20fd171
updated readme including contribution instructions
