### 2016-05-21 e1f9499
bump version number and update test hashes
