+++ paper.md
@@ -9,6 +9,7 @@ authors:
  - name: Matthew J. Wakefield
    orcid: 0000-0001-6624-4698
    affiliation: The Walter and Eliza Hall Institute
+   affiliation: The University of Melbourne
 date: 12 May 2016
 bibliography: paper.bib
 ---
