### 2016-05-21 #18 @genomematt
@xuanxu @afron - updated version addressing review comments pushed.  Examples are still a bit sparse but I will continue to add to them.  Data access issues with most of the worked examples I already have so I will need to hunt up some more datasets.  I really want these to be runnable notebooks. I also added an extra affiliation line to reflect my dual appointment to the paper.md
