### 2016-05-21 #18 @xuanxu
:+1: thanks @genomematt!
I've checked both missing points and I consider this publication-ready now.
A small doubt, @arfon & @genomematt: the [archive link](https://zenodo.org/record/51337) in this issue and the "DOI for current release" link in the "How to cite Xenomapper" section both still point to version 1.0.0. Probably a new DOI should be created for the current 1.0.1 version and have both links updated before paper is published.
