### 2016-05-22 5b18385
Update readme.md citing for JOSS
