### 2016-05-22 #18 @genomematt
Ok @arfon and @xuanxu I have made an updated release with a new zenodo id 10.5281/zenodo.51798 but this is in the whedon part of the submission not the paper.md so I can't see a way for me to edit it.

Updated the how to cite to reflect JOSS, and the Zenodo badge updates automatically (links to current version not a fixed DOI).
