### 2016-05-22 #18 @arfon
Thanks @xuanxu for the review.

> @arfon PDF will need regenerating.

👍 - yes, it's regenerated and should be correct at https://github.com/openjournals/joss-papers/blob/master/joss.00018/10.21105.joss.00018.pdf

@genomematt - your DOI _will_ be http://dx.doi.org/10.21105/joss.00018 but the Crossref system seems to be a little backed up right now (this paper is pretty far back in the queue).

> Where do I sign up as a reviewer and evangelist?

I'd be happy to add you to the `@openjournals/joss-reviewers` team if that sounds good?

:party:
