### 2016-05-22 #18 @genomematt
@arfon Only the second affiliation comes up on the PDF.  I will raise a bug.

Yes, add me to @openjournals/joss-reviewers
