### 2016-05-22 #18 @arfon
> Yes, add me to `@openjournals/joss-reviewers`

Excellent, thanks! You should have received an invite to join the team.
