### 2016-05-23 #18 @arfon
@genomematt - the DOI is now live: http://dx.doi.org/10.21105/joss.00018
